FileCache: make diagnostic rate conversion explicit
Fix the -Werror=double-promotion build failure introduced by a63df75aebfc. Preserve the diagnostic's double-precision arithmetic with an explicit cast.

Verified the original error and the corrected translation unit with the CE ARM compiler and original build flags using -fsyntax-only. Independent review passed; no full image build or device test.

Related: pannal/CoreELEC#79

+++ xbmc/filesystem/FileCache.cpp
@@ -319,8 +319,8 @@ void CFileCache::Process()
                   "throttle target {:.0f} B/s ({:.2f} Mbit/s), forward {} bytes "
                   "(nominal capacity {} bytes), average cache write rate {} B/s, throttle {}",
                   baseRate, readFactor, useAdaptativeReadFactor ? "adaptive" : "fixed", targetRate,
-                  targetRate * 8.0 / 1000000.0, forward, m_maxForward, m_writeRateActual,
-                  throttleState);
+                  static_cast<double>(targetRate) * 8.0 / 1000000.0, forward, m_maxForward,
+                  m_writeRateActual, throttleState);
         nextRateLog = now + 5s;
         lastLoggedBaseRate = baseRate;
       }
